+++ negocio/AnzueloNeg.cs
@@ -27,7 +27,7 @@ namespace negocio
                     aux.Numero = (string)datos.Lector["Numero"];
                     aux.Descripcion = (string)datos.Lector["Descripcion"];
                     aux.Cantidad = (int)datos.Lector["Cantidad"];
-                    aux.Peso = (float)datos.Lector["Peso"];
+                    aux.Peso = (double)datos.Lector["Peso"];
 
                     lista.Add(aux);
                 }
